feat: add e2e weekly distribution test and fix Node.js compat
- Add `src/config/env.js` helper to read env vars from both
  `import.meta.env` (Vite) and `process.env` (Node.js)
- Update `chains.js` and `contracts.js` to use `getEnv()` instead
  of `import.meta.env` directly
- Fix `walletDiscovery.js`: use `activeChain`, `DEPLOY_BLOCK`, and
  `RPC_URL` env var instead of hardcoded `base` chain and `fromBlock: 0n`
- Fix `createDistributions.js`: use `activeChain`, `RPC_URL`, and
  chain block timestamp instead of `Date.now()` for fork compatibility
- Fix `merkleDistributor.js` ABI: `amountPerClaim` is `uint176` and
  `walletCount` is `uint40` (not `uint256`) — wrong types caused
  selector mismatch and silent revert on Mint Club contract
- Add `scripts/e2e-test.sh`: full pipeline test on Anvil fork of
  Base Sepolia (7 check-ins → streak verification → wallet discovery
  → merkle tree → Pinata pin → Mint Club distribution)
- Add `contracts/src/DemoToken.sol`: minimal ERC-20 for testnet use
- Append fork-testing guide to `contracts/README.md`

+++ src/config/abis/merkleDistributor.js
@@ -6,8 +6,8 @@ export const merkleDistributorAbi = [
     inputs: [
       { name: "token", type: "address" },
       { name: "isERC20", type: "bool" },
-      { name: "amountPerClaim", type: "uint256" },
-      { name: "walletCount", type: "uint256" },
+      { name: "amountPerClaim", type: "uint176" },
+      { name: "walletCount", type: "uint40" },
       { name: "startTime", type: "uint40" },
       { name: "endTime", type: "uint40" },
       { name: "merkleRoot", type: "bytes32" },
